Add task Form Cleanly designed with a warning text when an empty task is added
